(bug) Fix Relative Paths
Fix an issue where the selected directory would be set to an absolute
path, causing the type to be generated at the current directory +
absolute path. Now, it will compare the selected path to the current
directory and create the appropriate relative path. For example, if
the current directory is '/src/app/feature/test/components' and the
selected directory is '/src/app/feature/other-test/services', the
resulting path would be '../../other-test/services'. This allows the
type to be generated in the correct directory.