(feat) Add Quick Open Schematic
Add the ability to open certain schematic types quickly using the
provided key-mapping. For example, to open a component file, use `C-c a
o c`. This will display a completing-read list of all components in the
project. Extended this functionality for all schematic types that have
a unique .schematic-type.ts pattern.